Use jsonwebtoken for authentication, add authorization middleware and checks, connecting user and posts

+++ restapi_project/controllers/feed.js
@@ -2,6 +2,7 @@ const fs = require('fs');
 const path = require('path');
 const { validationResult } = require('express-validator/check');
 const Post = require('../models/post');
+const User = require('../models/user');
 const errHandler = require('../misc/error-handler');
 
 exports.getPosts = (req, res, next) => {
@@ -45,23 +46,27 @@ exports.createPost = (req, res, next) => {
   const imageUrl = req.file.path;
   const title = req.body.title;
   const content = req.body.content;
-  // create post in db
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: {
-      name: 'Jimmy'
-    }
+    creator: req.userId
   });
 
   post
     .save()
     .then(result => {
-      res.status(201).json({
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      user.posts.push(post);
+      return user.save();
+    })
+    .then(user => {
+      return res.status(201).json({
         message: 'Post created successfully!',
-        post: result,
-        createdAt: new Date()
+        post: post,
+        creator: { _id: user._id, name: user.name }
       });
     })
     .catch(err => next(errHandler(err)));
@@ -115,6 +120,11 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized...');
+        error.statusCode = 403;
+        throw error;
+      }
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -142,12 +152,23 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized...');
+        error.statusCode = 403;
+        throw error;
+      }
       // check logged in user
       clearImage(post.imageUrl);
       return Post.findByIdAndDelete(postId);
     })
     .then(result => {
-      console.log(result);
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then(result => {
       res.status(200).json({ message: 'Deleted Post.' });
     })
     .catch(err => next(errHandler(err)));
